staking:
Comment out authorized sender validation for testing purposes, it might
deprecated in the future if having external callers to collect does not
impose any harm to the economic incentive structure.

The check was introduced in the first place when calling collect auto-settled
the allocation, this is not now the case.